keys_generator/requirements.txt
keyring==21.5.0
keyrings.alt==4.2.0
pycryptodome==3.17
pyinstaller==5.8.0
SecretStorage~=3.3.3
cryptography~=39.0.1